Misc updates
Remove safari hidden ability check for gen7 origin pkm
load template data on any save load, causes the legality triangle to
appear on supported games

+++ PKHeX.WinForms/Util/ConfigUtil.cs
@@ -15,7 +15,7 @@ namespace PKHeX.WinForms
             catch (ConfigurationErrorsException e)
             {
                 string path = (e.InnerException as ConfigurationErrorsException)?.Filename;
-                if (path != null)
+                if (path != null && File.Exists(path))
                     File.Delete(path);
                 return false;
             }
